fix response mapper id

+++ data/db/src/main/java/com/komeyama/simple/weather/db/DetailForecastEntity.kt
@@ -4,7 +4,7 @@ import com.komeyama.simple.weather.db.internal.entity.DetailImageEntityImpl
 import com.komeyama.simple.weather.db.internal.entity.TemperatureEntityImpl
 
 interface DetailForecastEntity {
-    var parentId: Long
+    var parentId: Int
     var date: String?
     var dateLabel: String?
     var telop: String?
